+++ docs/court_wizard.js
@@ -321,28 +321,28 @@ if (!('encodeInto' in cachedTextEncoder)) {
 
 let WASM_VECTOR_LEN = 0;
 
-function __wasm_bindgen_func_elem_201263(arg0, arg1, arg2, arg3) {
-    wasm.__wasm_bindgen_func_elem_201263(arg0, arg1, addHeapObject(arg2), addHeapObject(arg3));
+function __wasm_bindgen_func_elem_196725(arg0, arg1, arg2) {
+    wasm.__wasm_bindgen_func_elem_196725(arg0, arg1, addHeapObject(arg2));
 }
 
-function __wasm_bindgen_func_elem_201265(arg0, arg1, arg2) {
-    wasm.__wasm_bindgen_func_elem_201265(arg0, arg1, addHeapObject(arg2));
+function __wasm_bindgen_func_elem_201424(arg0, arg1, arg2) {
+    wasm.__wasm_bindgen_func_elem_201424(arg0, arg1, addHeapObject(arg2));
 }
 
-function __wasm_bindgen_func_elem_201277(arg0, arg1, arg2) {
-    wasm.__wasm_bindgen_func_elem_201277(arg0, arg1, isLikeNone(arg2) ? 0 : addHeapObject(arg2));
+function __wasm_bindgen_func_elem_201436(arg0, arg1, arg2) {
+    wasm.__wasm_bindgen_func_elem_201436(arg0, arg1, isLikeNone(arg2) ? 0 : addHeapObject(arg2));
 }
 
-function __wasm_bindgen_func_elem_194441(arg0, arg1) {
-    wasm.__wasm_bindgen_func_elem_194441(arg0, arg1);
+function __wasm_bindgen_func_elem_194600(arg0, arg1) {
+    wasm.__wasm_bindgen_func_elem_194600(arg0, arg1);
 }
 
-function __wasm_bindgen_func_elem_201267(arg0, arg1) {
-    wasm.__wasm_bindgen_func_elem_201267(arg0, arg1);
+function __wasm_bindgen_func_elem_201426(arg0, arg1) {
+    wasm.__wasm_bindgen_func_elem_201426(arg0, arg1);
 }
 
-function __wasm_bindgen_func_elem_196566(arg0, arg1, arg2) {
-    wasm.__wasm_bindgen_func_elem_196566(arg0, arg1, addHeapObject(arg2));
+function __wasm_bindgen_func_elem_201422(arg0, arg1, arg2, arg3) {
+    wasm.__wasm_bindgen_func_elem_201422(arg0, arg1, addHeapObject(arg2), addHeapObject(arg3));
 }
 
 const __wbindgen_enum_GamepadMappingType = ["", "standard"];
@@ -2187,24 +2187,14 @@ function __wbg_get_imports() {
         const ret = getObject(arg0).y;
         return ret;
     };
-    imports.wbg.__wbindgen_cast_02d4a10bf34c2625 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 156284, function: Function { arguments: [NamedExternref("FocusEvent")], shim_idx: 156287, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201064, __wasm_bindgen_func_elem_201265);
+    imports.wbg.__wbindgen_cast_127ff9300215f7a6 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 156365, function: Function { arguments: [NamedExternref("Array<any>"), NamedExternref("ResizeObserver")], shim_idx: 156366, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201223, __wasm_bindgen_func_elem_201422);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_068e556d15a36dab = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 156284, function: Function { arguments: [NamedExternref("Event")], shim_idx: 156287, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201064, __wasm_bindgen_func_elem_201265);
-        return addHeapObject(ret);
-    };
-    imports.wbg.__wbindgen_cast_095525ba18198245 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 156284, function: Function { arguments: [NamedExternref("PageTransitionEvent")], shim_idx: 156287, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201064, __wasm_bindgen_func_elem_201265);
-        return addHeapObject(ret);
-    };
-    imports.wbg.__wbindgen_cast_13069cca2b8f2d00 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 156284, function: Function { arguments: [Option(NamedExternref("Blob"))], shim_idx: 156313, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201064, __wasm_bindgen_func_elem_201277);
+    imports.wbg.__wbindgen_cast_134e92d6199c298a = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 156365, function: Function { arguments: [NamedExternref("WheelEvent")], shim_idx: 156368, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201223, __wasm_bindgen_func_elem_201424);
         return addHeapObject(ret);
     };
     imports.wbg.__wbindgen_cast_2241b6af4c4b2941 = function(arg0, arg1) {
@@ -2212,24 +2202,24 @@ function __wbg_get_imports() {
         const ret = getStringFromWasm0(arg0, arg1);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_263d2e7521c83802 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 156284, function: Function { arguments: [NamedExternref("WheelEvent")], shim_idx: 156287, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201064, __wasm_bindgen_func_elem_201265);
+    imports.wbg.__wbindgen_cast_5618f706e7b9f3f9 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 156365, function: Function { arguments: [NamedExternref("KeyboardEvent")], shim_idx: 156368, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201223, __wasm_bindgen_func_elem_201424);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_300dfe6813129f09 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 156284, function: Function { arguments: [NamedExternref("KeyboardEvent")], shim_idx: 156287, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201064, __wasm_bindgen_func_elem_201265);
+    imports.wbg.__wbindgen_cast_5c26bb6a1fb5130e = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 154480, function: Function { arguments: [Externref], shim_idx: 154481, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_196745, __wasm_bindgen_func_elem_196725);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_41a800bdb5190ac1 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 154399, function: Function { arguments: [Externref], shim_idx: 154400, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_196586, __wasm_bindgen_func_elem_196566);
+    imports.wbg.__wbindgen_cast_744d825575c8da0d = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 156365, function: Function { arguments: [Option(NamedExternref("Blob"))], shim_idx: 156394, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201223, __wasm_bindgen_func_elem_201436);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_67a2c0ccf72161f5 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 156284, function: Function { arguments: [], shim_idx: 156289, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201064, __wasm_bindgen_func_elem_201267);
+    imports.wbg.__wbindgen_cast_750cc71560d69633 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 153938, function: Function { arguments: [], shim_idx: 153939, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_194611, __wasm_bindgen_func_elem_194600);
         return addHeapObject(ret);
     };
     imports.wbg.__wbindgen_cast_7c316abdc43840a3 = function(arg0, arg1) {
@@ -2237,9 +2227,9 @@ function __wbg_get_imports() {
         const ret = getArrayU32FromWasm0(arg0, arg1);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_91aae7c1381741d4 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 153857, function: Function { arguments: [], shim_idx: 153858, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_194452, __wasm_bindgen_func_elem_194441);
+    imports.wbg.__wbindgen_cast_94918d3391128c82 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 156365, function: Function { arguments: [NamedExternref("PointerEvent")], shim_idx: 156368, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201223, __wasm_bindgen_func_elem_201424);
         return addHeapObject(ret);
     };
     imports.wbg.__wbindgen_cast_9575fb55a66c262b = function(arg0, arg1) {
@@ -2247,14 +2237,14 @@ function __wbg_get_imports() {
         const ret = getArrayI32FromWasm0(arg0, arg1);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_b006d52f624c3cce = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 156284, function: Function { arguments: [NamedExternref("Array<any>"), NamedExternref("ResizeObserver")], shim_idx: 156285, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201064, __wasm_bindgen_func_elem_201263);
+    imports.wbg.__wbindgen_cast_a23fbbc3f4ef480d = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 156365, function: Function { arguments: [NamedExternref("PageTransitionEvent")], shim_idx: 156368, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201223, __wasm_bindgen_func_elem_201424);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_bad3a5e03f439ef4 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 156284, function: Function { arguments: [NamedExternref("PointerEvent")], shim_idx: 156287, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201064, __wasm_bindgen_func_elem_201265);
+    imports.wbg.__wbindgen_cast_b1a2176f11b164ce = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 156365, function: Function { arguments: [NamedExternref("Event")], shim_idx: 156368, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201223, __wasm_bindgen_func_elem_201424);
         return addHeapObject(ret);
     };
     imports.wbg.__wbindgen_cast_bbb4883c6389f1de = function(arg0, arg1) {
@@ -2277,14 +2267,24 @@ function __wbg_get_imports() {
         const ret = arg0;
         return addHeapObject(ret);
     };
+    imports.wbg.__wbindgen_cast_da9b44b11fca564d = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 156365, function: Function { arguments: [], shim_idx: 156370, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201223, __wasm_bindgen_func_elem_201426);
+        return addHeapObject(ret);
+    };
     imports.wbg.__wbindgen_cast_e47ceb6027f5c92c = function(arg0, arg1) {
         // Cast intrinsic for `Ref(Slice(I16)) -> NamedExternref("Int16Array")`.
         const ret = getArrayI16FromWasm0(arg0, arg1);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_fb476d851e10efe9 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 156284, function: Function { arguments: [NamedExternref("Array<any>")], shim_idx: 156287, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201064, __wasm_bindgen_func_elem_201265);
+    imports.wbg.__wbindgen_cast_ee9a8992b30b40b9 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 156365, function: Function { arguments: [NamedExternref("Array<any>")], shim_idx: 156368, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201223, __wasm_bindgen_func_elem_201424);
+        return addHeapObject(ret);
+    };
+    imports.wbg.__wbindgen_cast_f4032c8c56be6a4a = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 156365, function: Function { arguments: [NamedExternref("FocusEvent")], shim_idx: 156368, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_201223, __wasm_bindgen_func_elem_201424);
         return addHeapObject(ret);
     };
     imports.wbg.__wbindgen_cast_feefb5fadd6457fd = function(arg0, arg1) {